backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
qrcode==7.4.2
Pillow==10.0.1
djangorestframework-simplejwt==5.3.0
django-templated-mail==1.1.1
requests==2.31.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-redis==5.3.0
cloudinary==1.41.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
